+++ test/harness/actions/DisputeTamperingActions.ts
@@ -21,28 +21,9 @@ export class DisputeTampering {
         dispute.input.disputeAuditingDataHash = hash("0x42");
     }
 
-    static createTamperTimeoutParticipant(
-        wrongParticipantAddress: string,
-        blockHeight: number
-    ) {
-        return (dispute: DisputeStruct): void => {
-            dispute.input.timeout.participant = wrongParticipantAddress;
-            dispute.input.timeout.blockHeight = blockHeight;
-        };
-    }
-
     static tamperDoubleFault(dispute: DisputeStruct): void {
         dispute.input.disputeAuditingDataHash = hash("0x42");
         dispute.input.latestStateSnapshotHash = hash("0x43");
-    }
-
-    static tamperInvalidStateProof(dispute: DisputeStruct): void {
-        dispute.input.latestStateSnapshotHash = hash("0x42");
-    }
-
-    static tamperInvalidStateProofWithCalldata(dispute: DisputeStruct): void {
-        DisputeTampering.tamperInvalidStateProof(dispute);
-        dispute.postedAuditingData = true;
     }
 
     static tamperPartialAuditing(dispute: DisputeStruct): void {
@@ -117,14 +98,16 @@ export class DisputeTamperingActions {
     stubConstructDispute(
         peerIndex: number,
         tamper: DisputeTamper,
-        options?: { autoRestore?: boolean }
+        options?: { autoRestore?: boolean; markMalicious?: boolean }
     ): void {
         const peer = this.harness.getPeer(peerIndex);
         const disputeManager: DisputeManager = peer.stateManager.disputeManager;
 
-        this.harness.contextApi.markMaliciousPeer({
-            maliciousPeerIndex: peerIndex
-        });
+        if (options?.markMalicious ?? true) {
+            this.harness.contextApi.markMaliciousPeer({
+                maliciousPeerIndex: peerIndex
+            });
+        }
         this.restoreConstructDispute(peerIndex);
 
         const originalConstructDispute =
